initial refactor
