+++ README.md
@@ -1,1 +1,6 @@
 # lane30
+
+# Testing
+
+1. Run `python3 -m http.server 8000 -d ./` from project directory.
+2. Open `http://[::]:8000/` in browser.
