+++ index.html
@@ -127,12 +127,7 @@
       </table>
     </div>
     <canvas id="lane" width="640" height="900"></canvas>
-    <!-- <script src="game.js"></script> -->
     <script src="main.js" type="module"></script>
-     
-    <!-- LOCAL TESTING -->
-    <!-- <script src="main.js"></script> -->
-    <!-- <script src="game_refactor.js"></script> -->
   </div>
 </body>
 </html>
