+++ main.js
@@ -1,4 +1,4 @@
-import { Game } from './game_refactor.js';
+import { Game } from './game.js';
 
 (() => {
   const game = new Game();
